Updates ContainerInterop to use PSR-11 (breaks BC)
This is a breaking change from the previous version, as the
[ContainerInterop][1] interfaces are no longer implemented. As of their
recommendation, but not because of it :wink:, we are now conforming the
[PSR-11][2] interfaces.

Not all interfaces were implemented (:eyes: Exceptions), so start
using them - conforming with previous exceptions thrown by
Respect\Config. Change 95ab990 modified an `\InvalidArgumentException`
to an `Exception` type, this reverts that change to conform with
previous contracts.

Rather than testing interoperability inside `ContainerTest` test case,
extract those tests into their own test case under a more friendly
namespace (`Test\Feature`) and better test names - along with the
test for the exception bullshit.

Technical debts
---------------

The test suite now holds two different organizations, one (introduced
here) unfinished. I've already dived into this code base a few times
and tests helped me a lot, but having them matching their concrete classes
might not be the best organization for that. As splitting this suite into
groups like "unit" and "integration" may not as useful as it is inside
a huge application (with dozens of dependencies), we can use something
like: "feature", "issues" and "library".

[1]: https://github.com/container-interop/container-interop
[2]: https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-11-container.md